tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/send/001.html

<!doctype html>
<title>WebSockets: send() with no args</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
  assert_throws(new TypeError(), function(){ws.send()});
});
</script>
